backend/requirements.txt
fastapi
uvicorn[standard]
websockets
python-socketio
pydantic
watchdog
psutil